requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.2

langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
langchain-core==0.3.18
langchain-ollama==0.2.0

psycopg2-binary==2.9.9
pymysql==1.1.0
pymongo==4.6.1
sqlalchemy==2.0.25

python-dotenv==1.0.1
openai==1.54.4
tiktoken==0.8.0
httpx==0.27.2
aiohttp==3.10.0

chromadb==0.4.22
onnxruntime==1.17.1
cryptography==42.0.5
